Cliques on mutigraph/directed graph types (#4502)
* updated functions to raise an exception for directed and multigraph

* added maximum_independent_set in the clique module and and updated tests

* updated docs

* module docstring
