+++ networkx/algorithms/approximation/tests/test_clique.py
@@ -1,13 +1,11 @@
-"""Unit tests for the :mod:`networkx.algorithms.approximation.clique`
-module.
-
-"""
+"""Unit tests for the :mod:`networkx.algorithms.approximation.clique` module."""
 
 
 import networkx as nx
 from networkx.algorithms.approximation import max_clique
 from networkx.algorithms.approximation import clique_removal
 from networkx.algorithms.approximation import large_clique_size
+from networkx.algorithms.approximation import maximum_independent_set
 
 
 def is_independent_set(G, nodes):
@@ -105,3 +103,9 @@ def test_large_clique_size():
     assert large_clique_size(G) == 8
     G.remove_edge(2, 3)
     assert large_clique_size(G) == 7
+
+
+def test_independent_set():
+    # smoke test
+    G = nx.Graph()
+    assert len(maximum_independent_set(G)) == 0
